+++ corpus/corpus2025-05-03-abcd1234.json
@@ -37,17 +37,17 @@
       "rightPinCount": 0,
       "topPinCount": 1,
       "bottomPinCount": 1,
-      "centerX": -2.4000000000000004,
+      "centerX": -3.2,
       "centerY": 1.7,
       "pins": [
         {
           "pinNumber": 2,
-          "x": -2.4000000000000004,
+          "x": -3.2,
           "y": 2.2
         },
         {
           "pinNumber": 1,
-          "x": -2.4000000000000004,
+          "x": -3.2,
           "y": 1.2
         }
       ]
@@ -58,8 +58,8 @@
       "netId": "NET1",
       "netLabelId": "loaded-nl-e89d34ed-c264-4102-a8d1-c9610c6261a7-0",
       "anchorPosition": "left",
-      "x": -2.4000000000000004,
-      "y": 0.6000000000000001
+      "x": -3.2,
+      "y": 0.4
     },
     {
       "netId": "NET2",
@@ -73,7 +73,7 @@
     {
       "points": [
         {
-          "x": -2.4000000000000004,
+          "x": -3.2,
           "y": 2.2
         },
         {
@@ -97,12 +97,12 @@
     {
       "points": [
         {
-          "x": -2.4000000000000004,
+          "x": -3.2,
           "y": 1.2
         },
         {
-          "x": -2.4000000000000004,
-          "y": 0.6000000000000001
+          "x": -3.2,
+          "y": 0.4
         }
       ],
       "from": {
